Add OpenMP directives for expensive calculations
With a sleep of 1 s, NINDIVIDUALS=4, and ten generations max for each
individual, we should spend a total of ~40 s to finish the program.
With num of jobs=4, this should reduce to 10 s which is indeed the
case.

* pso.f90 (pso, fresh_individuals): Add OpenMP directives to calculate
the score for each individual in parallel.
(score_gen): Add sleep call to emulate expensive calculation.